Reorganize STM32CubeIDE integration
