+++ Firmware/CoverUI/CoverUI/platformio.ini
@@ -41,8 +41,12 @@ build_src_filter =
   -<LEDcontrol.cpp>
   -<statemachine.cpp>
   +<yfc500/*>
+  -<yfc500/stm32cube/src/main.c>
   
-build_flags = ${env.build_flags} -DHW_YFC500
+build_flags =
+  ${env.build_flags}
+  -I yfc500/stm32cube/inc
+  -DHW_YFC500
 
 ; Need at least openocd 0.11 for working CMSIS-DAP adapter detection
 platform_packages = platformio/tool-openocd@^2.1100.211028
@@ -74,10 +78,13 @@ build_src_filter =
   -<LEDcontrol.cpp>
   -<statemachine.cpp>
   +<yfc500/*>
-  ; Use Semihosting's syscall instead of ours
-  -<yfc500/stm32cube/syscalls.c>
+  -<yfc500/stm32cube/src/main.c>
+  -<yfc500/stm32cube/src/syscalls.c> ; Use Semihosting's syscall instead of ours
 
-build_flags = ${env.build_flags} -DHW_YFC500
+build_flags =
+  ${env.build_flags}
+  -I yfc500/stm32cube/inc
+  -DHW_YFC500
 
 ; Semihosting see https://piolabs.com/blog/insights/debugging-semihosting.html
 build_unflags =
