+++ Firmware/CoverUI/CoverUI/yfc500/stm32cube/README.txt
@@ -1,0 +1,3 @@
+STM32CubeIDE files.
+Simply copied from STM32CubeIDE's 'Src' and 'Inc' folders,
+or symlink these into Eclipse's STM32CubeIDE's workspace.
