+++ Firmware/CoverUI/CoverUI/yfc500/stm32cube/YardForce Classic 500 STM32F030R8Tx.ioc
@@ -1,0 +1,234 @@
+#MicroXplorer Configuration settings - do not modify
+CAD.formats=
+CAD.pinconfig=
+CAD.provider=
+Dma.Request0=USART2_RX
+Dma.Request1=USART2_TX
+Dma.RequestsNb=2
+Dma.USART2_RX.0.Direction=DMA_PERIPH_TO_MEMORY
+Dma.USART2_RX.0.Instance=DMA1_Channel5
+Dma.USART2_RX.0.MemDataAlignment=DMA_MDATAALIGN_BYTE
+Dma.USART2_RX.0.MemInc=DMA_MINC_ENABLE
+Dma.USART2_RX.0.Mode=DMA_NORMAL
+Dma.USART2_RX.0.PeriphDataAlignment=DMA_PDATAALIGN_BYTE
+Dma.USART2_RX.0.PeriphInc=DMA_PINC_DISABLE
+Dma.USART2_RX.0.Priority=DMA_PRIORITY_LOW
+Dma.USART2_RX.0.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlignment,MemDataAlignment,Mode,Priority
+Dma.USART2_TX.1.Direction=DMA_MEMORY_TO_PERIPH
+Dma.USART2_TX.1.Instance=DMA1_Channel4
+Dma.USART2_TX.1.MemDataAlignment=DMA_MDATAALIGN_BYTE
+Dma.USART2_TX.1.MemInc=DMA_MINC_ENABLE
+Dma.USART2_TX.1.Mode=DMA_NORMAL
+Dma.USART2_TX.1.PeriphDataAlignment=DMA_PDATAALIGN_BYTE
+Dma.USART2_TX.1.PeriphInc=DMA_PINC_DISABLE
+Dma.USART2_TX.1.Priority=DMA_PRIORITY_LOW
+Dma.USART2_TX.1.RequestParameters=Instance,Direction,PeriphInc,MemInc,PeriphDataAlignment,MemDataAlignment,Mode,Priority
+File.Version=6
+GPIO.groupedBy=Group By Peripherals
+KeepUserPlacement=false
+Mcu.CPN=STM32F030R8T6
+Mcu.Family=STM32F0
+Mcu.IP0=DMA
+Mcu.IP1=NVIC
+Mcu.IP2=RCC
+Mcu.IP3=TIM16
+Mcu.IP4=TIM17
+Mcu.IP5=USART1
+Mcu.IP6=USART2
+Mcu.IPNb=7
+Mcu.Name=STM32F030R8Tx
+Mcu.Package=LQFP64
+Mcu.Pin0=PC0
+Mcu.Pin1=PC1
+Mcu.Pin10=PA6
+Mcu.Pin11=PA7
+Mcu.Pin12=PC4
+Mcu.Pin13=PB0
+Mcu.Pin14=PA9
+Mcu.Pin15=PA10
+Mcu.Pin16=PA14
+Mcu.Pin17=PA15
+Mcu.Pin18=PC10
+Mcu.Pin19=PC11
+Mcu.Pin2=PC2
+Mcu.Pin20=PC12
+Mcu.Pin21=PD2
+Mcu.Pin22=PB3
+Mcu.Pin23=PB4
+Mcu.Pin24=VP_TIM16_VS_ClockSourceINT
+Mcu.Pin25=VP_TIM17_VS_ClockSourceINT
+Mcu.Pin3=PC3
+Mcu.Pin4=PA0
+Mcu.Pin5=PA1
+Mcu.Pin6=PA2
+Mcu.Pin7=PA3
+Mcu.Pin8=PA4
+Mcu.Pin9=PA5
+Mcu.PinsNb=26
+Mcu.ThirdPartyNb=0
+Mcu.UserConstants=TIM_BLINK_SLOW,TIM16;TIM_BLINK_FAST,TIM17;UART_LL,USART2;UART_U4,USART1
+Mcu.UserName=STM32F030R8Tx
+MxCube.Version=6.8.0
+MxDb.Version=DB.6.0.80
+NVIC.DMA1_Channel4_5_IRQn=true\:0\:0\:false\:false\:true\:false\:true\:true
+NVIC.ForceEnableDMAVector=true
+NVIC.HardFault_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
+NVIC.NonMaskableInt_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
+NVIC.PendSV_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:false
+NVIC.SVC_IRQn=true\:0\:0\:false\:false\:true\:false\:false\:true
+NVIC.SysTick_IRQn=true\:3\:0\:false\:false\:true\:false\:true\:false
+NVIC.TIM16_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
+NVIC.TIM17_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
+NVIC.USART1_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
+NVIC.USART2_IRQn=true\:0\:0\:false\:false\:true\:true\:true\:true
+PA0.GPIOParameters=GPIO_Speed,GPIO_Label,GPIO_ModeDefaultOutputPP
+PA0.GPIO_Label=LED_S1
+PA0.GPIO_ModeDefaultOutputPP=GPIO_MODE_OUTPUT_PP
+PA0.GPIO_Speed=GPIO_SPEED_FREQ_LOW
+PA0.Locked=true
+PA0.Signal=GPIO_Output
+PA1.GPIOParameters=GPIO_Label
+PA1.GPIO_Label=LED_S2
+PA1.Locked=true
+PA1.Signal=GPIO_Output
+PA10.Locked=true
+PA10.Mode=Asynchronous
+PA10.Signal=USART1_RX
+PA14.Locked=true
+PA14.Signal=SYS_SWCLK
+PA15.GPIOParameters=GPIO_Label
+PA15.GPIO_Label=LED_MON
+PA15.Locked=true
+PA15.Signal=GPIO_Output
+PA2.GPIOParameters=GPIO_Label
+PA2.GPIO_Label=UART_LL_TX
+PA2.Locked=true
+PA2.Mode=Asynchronous
+PA2.Signal=USART2_TX
+PA3.GPIOParameters=GPIO_Label
+PA3.GPIO_Label=UART_LL_RX
+PA3.Mode=Asynchronous
+PA3.Signal=USART2_RX
+PA4.GPIOParameters=GPIO_Label
+PA4.GPIO_Label=LED_2HR
+PA4.Locked=true
+PA4.Signal=GPIO_Output
+PA5.GPIOParameters=GPIO_Label
+PA5.GPIO_Label=LED_4HR
+PA5.Locked=true
+PA5.Signal=GPIO_Output
+PA6.GPIOParameters=GPIO_Label
+PA6.GPIO_Label=LED_6HR
+PA6.Locked=true
+PA6.Signal=GPIO_Output
+PA7.GPIOParameters=GPIO_Label
+PA7.GPIO_Label=LED_8HR
+PA7.Locked=true
+PA7.Signal=GPIO_Output
+PA9.Locked=true
+PA9.Mode=Asynchronous
+PA9.Signal=USART1_TX
+PB0.GPIOParameters=GPIO_Label
+PB0.GPIO_Label=LED_SMD
+PB0.Locked=true
+PB0.Signal=GPIO_Output
+PB3.GPIOParameters=GPIO_Label
+PB3.GPIO_Label=LED_SAT
+PB3.Locked=true
+PB3.Signal=GPIO_Output
+PB4.GPIOParameters=GPIO_Label
+PB4.GPIO_Label=LED_SUN
+PB4.Locked=true
+PB4.Signal=GPIO_Output
+PC0.GPIOParameters=GPIO_Label
+PC0.GPIO_Label=LED_LIFTED
+PC0.Locked=true
+PC0.Signal=GPIO_Output
+PC1.GPIOParameters=GPIO_Label
+PC1.GPIO_Label=LED_WIRE
+PC1.Locked=true
+PC1.Signal=GPIO_Output
+PC10.GPIOParameters=GPIO_Label
+PC10.GPIO_Label=LED_TUE
+PC10.Locked=true
+PC10.Signal=GPIO_Output
+PC11.GPIOParameters=GPIO_Label
+PC11.GPIO_Label=LED_WED
+PC11.Locked=true
+PC11.Signal=GPIO_Output
+PC12.GPIOParameters=GPIO_Label
+PC12.GPIO_Label=LED_THU
+PC12.Locked=true
+PC12.Signal=GPIO_Output
+PC2.GPIOParameters=GPIO_Label
+PC2.GPIO_Label=LED_BAT
+PC2.Locked=true
+PC2.Signal=GPIO_Output
+PC3.GPIOParameters=GPIO_Label
+PC3.GPIO_Label=LED_CHARGE
+PC3.Locked=true
+PC3.Signal=GPIO_Output
+PC4.GPIOParameters=GPIO_Label
+PC4.GPIO_Label=LED_LOCK
+PC4.Locked=true
+PC4.Signal=GPIO_Output
+PD2.GPIOParameters=GPIO_Label
+PD2.GPIO_Label=LED_FRI
+PD2.Locked=true
+PD2.Signal=GPIO_Output
+PinOutPanel.RotationAngle=0
+ProjectManager.AskForMigrate=true
+ProjectManager.BackupPrevious=true
+ProjectManager.CompilerOptimize=6
+ProjectManager.ComputerToolchain=false
+ProjectManager.CoupleFile=true
+ProjectManager.CustomerFirmwarePackage=
+ProjectManager.DefaultFWLocation=true
+ProjectManager.DeletePrevious=true
+ProjectManager.DeviceId=STM32F030R8Tx
+ProjectManager.FirmwarePackage=STM32Cube FW_F0 V1.11.4
+ProjectManager.FreePins=false
+ProjectManager.HalAssertFull=false
+ProjectManager.HeapSize=0x200
+ProjectManager.KeepUserCode=false
+ProjectManager.LastFirmware=true
+ProjectManager.LibraryCopy=1
+ProjectManager.MainLocation=Core/Src
+ProjectManager.NoMain=false
+ProjectManager.PreviousToolchain=
+ProjectManager.ProjectBuild=false
+ProjectManager.ProjectFileName=YardForce Classic 500 STM32 CoverUI.ioc
+ProjectManager.ProjectName=YardForce Classic 500 STM32 CoverUI
+ProjectManager.ProjectStructure=
+ProjectManager.RegisterCallBack=
+ProjectManager.StackSize=0x400
+ProjectManager.TargetToolchain=STM32CubeIDE
+ProjectManager.TemplateDestinationPath=/home/jebeling/STM32CubeIDE/workspace_1.12.0/YardForce Classic 500 STM32 CoverUI
+ProjectManager.TemplateSourcePath=/opt/st/stm32cubeide_1.12.0/plugins/com.st.stm32cube.common.mx_6.8.0.202302231600/db/extra_templates
+ProjectManager.ToolChainLocation=
+ProjectManager.UnderRoot=true
+ProjectManager.functionlistsort=1-SystemClock_Config-RCC-false-HAL-false,2-MX_GPIO_Init-GPIO-false-HAL-true,3-MX_DMA_Init-DMA-false-HAL-true,4-MX_USART2_UART_Init-USART2-false-HAL-true,5-MX_USART1_UART_Init-USART1-false-HAL-true,6-MX_TIM16_Init-TIM16-false-HAL-true,7-MX_TIM17_Init-TIM17-false-HAL-true
+RCC.FamilyName=M
+RCC.IPParameters=FamilyName,PLLCLKFreq_Value,PLLMCOFreq_Value,TimSysFreq_Value
+RCC.PLLCLKFreq_Value=8000000
+RCC.PLLMCOFreq_Value=4000000
+RCC.TimSysFreq_Value=8000000
+TIM16.IPParameters=Prescaler,Period
+TIM16.Period=499
+TIM16.Prescaler=7999
+TIM17.IPParameters=Prescaler,Period
+TIM17.Period=99
+TIM17.Prescaler=7999
+USART1.BaudRate=115200
+USART1.IPParameters=VirtualMode-Asynchronous,BaudRate
+USART1.VirtualMode-Asynchronous=VM_ASYNC
+USART2.BaudRate=115200
+USART2.IPParameters=VirtualMode-Asynchronous,BaudRate,OverrunDisableParam
+USART2.OverrunDisableParam=ADVFEATURE_OVERRUN_DISABLE
+USART2.VirtualMode-Asynchronous=VM_ASYNC
+VP_TIM16_VS_ClockSourceINT.Mode=Enable_Timer
+VP_TIM16_VS_ClockSourceINT.Signal=TIM16_VS_ClockSourceINT
+VP_TIM17_VS_ClockSourceINT.Mode=Enable_Timer
+VP_TIM17_VS_ClockSourceINT.Signal=TIM17_VS_ClockSourceINT
+board=custom
+isbadioc=false
