+++ Firmware/CoverUI/CoverUI/yfc500/stm32cube/inc/stm32f0xx_hal_conf.h
